Fix Inkquest Entity

+++ src/main/java/com/springcore/ai/scaiplatform/dto/InkQuest/InkCumulativePoint.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @Data
 public class InkCumulativePoint {
     private String month;
-    private Long words;
+    private Integer words;
 }
